Format code with goreturns
This brings in a lot of deltas, but I am not sure if the code was even
formatted before, at least for the changes brought by the lastest PR,
as there are even semicolons lying around.

Doing this as a separate commit in order to make reviewing a bit easier.